+++ algodot/src/algod/mod.rs
@@ -1,3 +1,391 @@
+pub mod bar {
+    //use algonaut_abi::abi_interactions::AbiMethodArg;
+    //use algonaut_abi::abi_interactions::AbiReturn;
+    use algonaut::abi::abi_type::AbiType;
+
+
+    use algonaut::abi::abi_interactions::AbiMethod;
+
+    pub struct Foo {
+        pub name: String,
+        pub description: String,
+        pub type_: String, 
+        pub parsed: Option<String>,
+    }
+
+    impl MyTrait for Foo {
+        type Foo = Foo;
+        type Type = String;
+        type Parsed = Option<String>;
+
+        fn new() -> Self::Foo {
+            Foo {
+                name: "".to_string(),
+                description: "".to_string(),
+                type_: "".to_string(),
+                parsed: None,
+            }
+        }
+
+        fn r#type() -> String { "".to_string() }
+        fn parsed() -> Option<AbiType> { None }
+    }
+
+    trait MyTrait {
+        type Foo;
+        type Type: ToString;
+        type Parsed;
+
+        fn new() -> Self::Foo;
+        fn r#type() -> String;
+        fn parsed() -> Option<AbiType>;
+    }
+
+    impl Foo {
+        //Doc : https://developer.algorand.org/docs/get-details/transactions/signatures/#single-signatures
+        //      https://developer.algorand.org/docs/get-details/dapps/smart-contracts/ABI/?from_query=Method%20Signature#reference-types
+        // Boilerplate
+        //pub fn new() -> AbiMethod {
+        //    let method_sig : String = "withdraw(uint64,account)void".to_string();
+            //let method_sig : String = "add(uint64,uint64)uint128".to_string();
+
+            
+        //    println!("{}",&method_sig);
+
+        //    AbiMethod::from_signature(&method_sig)
+        //    .expect("Error")
+            
+        //}
+        
+        pub fn withdraw() -> AbiMethod {
+            let method_sig : String = "withdraw(uint64,account)void".to_string();
+            //let method_sig : String = "add(uint64,uint64)uint128".to_string();
+
+            
+            println!("Method Signature: {}",&method_sig);
+
+            AbiMethod::from_signature(&method_sig)
+            .expect("Error")
+            
+        }
+
+        pub fn deposit() -> AbiMethod {
+            let method_sig : String = "deposit(PaymentTransaction,account)void".to_string();
+            //let method_sig : String = "add(uint64,uint64)uint128".to_string();
+
+            
+            println!("Method Signature: {}",&method_sig);
+
+            AbiMethod::from_signature(&method_sig)
+            .expect("Error")
+            
+        }
+   
+    }
+}
+
+
+pub mod escrow {
+
+    /*Atomic Transaction Composer Helper Modules*/
+    use algonaut::atomic_transaction_composer::transaction_signer::TransactionSigner::BasicAccount;
+    //use crate::core::Account;
+    use algonaut::transaction::account::Account;
+    //use crate::algod::Account;
+    
+    use algonaut::algod::v2::Algod;
+    use algonaut::abi::abi_type::AbiValue::Int;
+    use algonaut::core::Address;
+    
+    //use num_bigint::BigUint;
+
+    use algonaut::{
+        
+        atomic_transaction_composer::{
+            transaction_signer::TransactionSigner, AbiArgValue, //AbiMethodReturnValue,
+            AtomicTransactionComposer, //AbiReturnDecodeError, AddMethodCallParams, 
+            TransactionWithSigner, //AtomicTransactionComposerStatus, 
+        },
+        error::ServiceError,
+    };
+    
+   //;
+         //AbiArgType,AbiReturn,, ReferenceArgType,, AbiReturnType  
+        //abi_type::{AbiType, AbiValue as OtherAbiValue},
+    //};
+    use algonaut::core::{to_app_address, Address as OtherAddress, MicroAlgos, CompiledTeal};
+    use algonaut::abi::abi_interactions::AbiMethod;
+    use algonaut::transaction::{
+        builder::TxnFee, builder::TxnFee::Fixed,
+        transaction::{ApplicationCallOnComplete, StateSchema},
+        Pay, TxnBuilder,
+    };
+
+    use algonaut::core::SuggestedTransactionParams as OtherSuggestedTransactionParams;
+    use algonaut::transaction::{transaction::Payment}; //account::Account
+  
+    use algonaut::crypto::HashDigest;
+  
+    use std::convert::TryInto;
+   
+    //use crate::params::params::MySuggestedTransactionParams;
+    use gdnative::prelude::*;//::ToVariant;
+
+
+
+    #[derive(Clone, Debug, ToVariant)] //PartialEq,
+    //lifetime Parameter
+    pub struct Foo <'a> {
+        pub withdrw_amt: u32,
+        pub withdrw_to_addr: [u8; 32],
+        pub arg1: AbiArgValue,
+        pub arg2: AbiArgValue,
+        pub _app_id: u64,
+        pub _escrow_address: Address,
+        pub atc: &'a AtomicTransactionComposer,
+    }
+
+    trait MyTrait {
+        type Foo <'a>;
+        type Params;
+        type Parsed;
+        type Payment;
+
+        fn _app_id(&self, x: u64) -> u64;
+        //fn default() -> Option<String>{ None }
+        //fn suggested_tx_params(&self) -> OtherSuggestedTransactionParams { OtherSuggestedTransactionParams::default() }
+        fn arg1(withdrw_amt: u64) -> AbiArgValue { AbiArgValue::AbiValue(Int(withdrw_amt.into())) }
+        fn arg2(withdrw_amt: u64) -> AbiArgValue { AbiArgValue::AbiValue(Int(withdrw_amt.into())) }
+    }
+    //Docs: https://godot-rust.github.io/docs/gdnative/prelude/struct.Variant.html
+    impl ToVariant for  withdrw_to_addr { //[u8; 32]
+    fn to_variant(&self) -> PoolArray<u8> {
+        PoolArray::from_variant_array(self)
+       }
+    }
+
+    impl ToVariant for atc {
+    fn to_variant(&self) -> Variant {
+        Variant::new(self)
+    }
+
+    impl ToVariant for &&AtomicTransactionComposer {//: [u8; 32] {
+    fn to_variant(&self) -> Variant {
+        Variant::new(self)
+        }
+
+    }
+}
+
+    impl MyTrait for Foo <'_>{
+        type Foo <'a> = Foo<'a>;
+        type Parsed = Option<String>;
+        type Payment = Option<Payment>;
+        type Params = Option<OtherSuggestedTransactionParams>;
+        fn _app_id(&self, x: u64) -> u64 { x }
+    }
+
+    impl Foo <'_> {
+        // Adding method to create application call
+        fn get_call(&self) -> Result<ApplicationCallOnComplete, ServiceError> {
+            //let func_args = vec![self.arg1.clone(), self.arg2.clone()];
+            
+            todo!()
+            
+        }
+
+        // Adding method to create pay transaction
+        fn get_payment(&self) -> Result<Payment, ServiceError> {
+            todo!()
+           // tx
+        }
+
+        fn arg1(&self)-> AbiArgValue{ 
+            todo!()
+            
+        }
+        
+        pub fn note(size : u32) -> Option <Vec<u8>>{
+            Some(vec![size.try_into().unwrap()])
+
+        }
+    
+
+
+        pub fn withdraw_amount(amount : u32) -> AbiArgValue {
+            /*
+            Converts a U64 int to Big Uint and returns an AbiArg Value
+            
+            Temporarily Disabling
+            */
+            //let withdrw_amt : BigUint = BigUint::new(vec![amount]); //in MicroAlgos
+            
+
+            //let arg1 : AbiArgValue = AbiArgValue::AbiValue( Int(withdrw_amt));
+            //arg1
+
+
+            todo!()
+
+        }
+            
+    
+        
+        pub fn withdraw(_acct1: Account ){
+             /* 
+            Withdraw Method Parameters for Escrow SmartContract
+            
+                Docs: https://docs.rs/num-bigint/0.4.3/num_bigint/struct.BigUint.html
+
+                Does nothing
+            */
+
+        }
+        
+
+        //use algonaut_core::Address;
+        pub fn pay(to_address : algonaut::core::Address , acct1 : Account, _params : algonaut::core::SuggestedTransactionParams) -> algonaut::transaction::transaction::Transaction{
+            /*
+                Constructs a Payment Transaction to an Address
+            */
+
+             let _t = TxnBuilder::with(
+
+                    &_params,
+
+                    Pay::new(acct1.address(), to_address, MicroAlgos(123_456)).build(),
+
+                )
+
+                .build()
+                .unwrap();
+            
+            return _t;
+        }
+
+        pub fn app_address (app_id : &u64) -> Address{
+            to_app_address(*app_id)
+        }
+        
+        pub fn deposit(_algod : Algod , acct1_3 : Account ,  params : algonaut::core::SuggestedTransactionParams) -> algonaut::core::SuggestedTransactionParams {
+            /*
+            Deposit Method Parameters for Escrow SmartContract
+            Unused and Depreciated
+           
+            Does
+            */
+
+            //Params
+            //
+            //App ID
+            let _app_id = 161737986;
+
+            
+            //Get Escrow Address From App ID
+
+            let _escrow_address = Foo::app_address(&_app_id); //to_app_address(_app_id.clone());
+           
+            println!(" building Pay transaction to Escrow Address: {}", &_escrow_address);
+
+            let _t = Foo::pay(_escrow_address, acct1_3.clone().into(), params.clone());                
+
+            // create a transaction with signer with the current transaction
+
+            let _signer ;//= TransactionSigner::BasicAccount(acct1_3);
+
+
+            let tx_with_signer = TransactionWithSigner { tx: _t, signer: _signer };
+
+
+            let mut atc = AtomicTransactionComposer::default();  
+
+            // Deposit
+            // Add Payment Txn to 
+            // Should Ideally Match To A Statemachine Behaviour Bloc
+            atc.add_transaction(tx_with_signer).unwrap();
+
+            params
+
+ 
+        }
+
+        pub fn new() -> AtomicTransactionComposer{
+        /*
+        Constructs a Default Atomic Transation Composer
+        */
+            AtomicTransactionComposer::default()
+        
+        }
+     
+        pub fn address_to_bytes(addr: String) -> [u8; 32]{ 
+        /*
+        Constructs a 32 Bit Byte Slice froma Given Address String
+        */   
+            let mut _to_addr: [u8; 32] = [0; 32];
+            //_to_addr.copy_from_slice(&acct1.address().to_string().as_bytes()[..32]);
+            _to_addr.copy_from_slice(&addr.as_bytes()[..32]);
+
+            _to_addr
+            
+        }
+
+        //let arg2: AbiArgValue = AbiArgValue::AbiValue(algonaut_abi::abi_type::AbiValue::Address(OtherAddress::new(withdrw_to_addr)));
+      
+        pub fn address(addr : [u8; 32]) -> AbiArgValue {
+            /* Returns an Address abi value from an Address as [u8,32]*/
+            AbiArgValue::AbiValue(algonaut::abi::abi_type::AbiValue::Address(OtherAddress::new(addr)))
+
+        } 
+
+        
+        pub fn basic_account(mnemonic : &str)  ->  algonaut::atomic_transaction_composer::transaction_signer::TransactionSigner{
+            BasicAccount(Account::from_mnemonic(&mnemonic).unwrap())
+        
+        }
+
+        pub fn fee(amount : u64) -> TxnFee{Fixed(MicroAlgos(amount))}
+
+        pub fn construct_app_call_method(
+        /*
+        Constructs an App Call Method as a Rust Module
+        
+        */
+        
+        //&AtomicTransactionComposer
+        &self,
+        //#[base] _base: &Node,
+        _app_id: u64,
+        _method: AbiMethod,
+        _method_args: Vec<AbiArgValue>,
+        _fee: TxnFee,//Fixed(MicroAlgos(2500u64)), //make customizable
+        _sender: Address,
+        _on_complete: ApplicationCallOnComplete,
+        _clear_program: Option<CompiledTeal>,
+        _global_schema: Option<StateSchema>,
+        _local_schema: Option<StateSchema>,
+        _extra_pages: u32,
+        _note: Option<Vec<u8>>,
+        _lease: Option<HashDigest>,
+        _rekey_to: Option<Address>,
+        _signer: TransactionSigner,
+    
+    
+    ) -> Result<Foo<'_>, ServiceError> {
+            todo!()
+            
+        }
+        
+
+    } 
+
+}
+
+
+
+
+
+
+
 use algodot_core::*;
 use algodot_macros::*;
 use algonaut::algod::v2::Algod;
@@ -13,6 +401,17 @@ use gdnative::prelude::*;
 use gdnative::tasks::{Async, AsyncMethod, Spawner};
 use std::rc::Rc;
 
+
+use algonaut::atomic_transaction_composer::{AtomicTransactionComposerStatus, 
+AddMethodCallParams //transaction_signer::TransactionSigner::BasicAccount, 
+};
+
+//cant get my mods
+//use super::escrow::Foo;
+use crate::algod::escrow::Foo;
+use bar::Foo as OtherFoo;
+use algodot_core::Account;
+use algonaut::transaction::transaction::ApplicationCallOnComplete::NoOp;
 
 #[derive(NativeClass)]
 #[inherit(Node)]
@@ -296,6 +695,78 @@ impl Algodot {
     }
 
     #[method]
+    #[allow(clippy::too_many_arguments)]
+    async fn construct_atc(
+        /* Atomic Transaction Composer*/
+        &self,
+        #[base] _base: &Node,
+        params: SuggestedTransactionParams,
+        sender: Address,
+        mnemonic : String,
+        app_id: u64,
+        #[opt] app_arguments: Option<String>, 
+        
+   
+    ) -> Result<(), Foo> { 
+
+       
+    let mut atc = escrow::Foo::new();  
+
+
+
+    let mut _to_addr: [u8; 32] = Foo::address_to_bytes(sender.to_string());//[0; 32];
+
+    let __app_id : u64 = 161737986 ;
+    let pages: u32 = 0;
+    
+    godot_dbg!("retrieving suggested params");
+    let params = self.algod.suggested_transaction_params().await.unwrap();
+    //Txn Details As a Struct
+    let details = escrow::Foo { 
+            withdrw_amt : 0u32,//Foo::withdraw_amount(0u32),//BigUint::new(vec![0]),//BigUint { data: vec![0u64] },//BigUint = BigUint::new(vec![0]), 
+            withdrw_to_addr: _to_addr.clone(), 
+            arg1: Foo::withdraw_amount(0u32), 
+            arg2: Foo::address(_to_addr),
+            _app_id: __app_id.clone(), 
+            _escrow_address: Foo::app_address(&__app_id),//to_app_address(__app_id), 
+            atc: &atc };
+
+    //println!("{:?}", &details);
+    godot_dbg!(&details);
+            //Add method Call     
+    atc.add_method_call( &mut AddMethodCallParams {
+                    app_id: details._app_id,
+                    method: bar::Foo::withdraw(), //bar::Foo::withdraw() //for deposits //bar::Foo::deposit()
+                    method_args: vec![details.arg1, details.arg2],
+                    fee: escrow::Foo::fee(2500),
+                    sender: *sender,
+                    suggested_params: params,
+                    on_complete: NoOp,
+                    approval_program: None,
+                    clear_program: None,
+                    global_schema: None,
+                    local_schema: None,
+                    extra_pages: pages,
+                    note: Foo::note(0u32),//_note,
+                    lease: None,
+                    rekey_to: None,
+                    signer: Foo::basic_account(&mnemonic)
+            
+        }
+    ).unwrap();
+
+
+    atc.build_group().expect("Error");
+
+    atc.execute(&self.algod).await.expect("Error");
+    
+    let status_str : &mut AtomicTransactionComposerStatus = &mut atc.status();
+    godot_dbg!(status_str);
+
+    Ok(())
+    }
+
+    #[method]
     fn construct_asset_opt_in(
         &self,
         #[base] _base: &Node,
@@ -408,4 +879,7 @@ asyncmethods!(algod, node, this,
             godot_unwrap!(compiled).map(|c| (c.hash().0.to_vec().to_variant(), c.bytes_to_sign().to_variant())).to_variant()
         }
     }
+
+    //fn execute(&algod) {atc.execute(&algod).await.expect("Error")
+    //}
 );
